#122 [feat] OperationInfoModel.kt 구현 [fix] 패키지명 변경
