+++ app/src/main/java/kr/co/lion/application/finalproject_aparttalk/ui/entiremenu/OperationInfo/fragment/FinancialSystemFragment.kt
@@ -1,4 +1,4 @@
-package kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptOperationInfo.fragment
+package kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.OperationInfo.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,25 +7,24 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.google.android.material.divider.MaterialDividerItemDecoration
-import kr.co.lion.application.finalproject_aparttalk.R
-import kr.co.lion.application.finalproject_aparttalk.databinding.FragmentSafetyManagementBinding
-import kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptOperationInfo.adapter.OperationSecondRecyclerView
+import kr.co.lion.application.finalproject_aparttalk.databinding.FragmentFinancialSystemBinding
+import kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.OperationInfo.adapter.OperationSecondRecyclerView
 
-class SafetyManagementFragment : Fragment() {
+class FinancialSystemFragment : Fragment() {
 
-    lateinit var fragmentSafetyManagementBinding: FragmentSafetyManagementBinding
+    lateinit var fragmentFinancialSystemBinding: FragmentFinancialSystemBinding
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         // Inflate the layout for this fragment
-        fragmentSafetyManagementBinding = FragmentSafetyManagementBinding.inflate(layoutInflater)
+        fragmentFinancialSystemBinding = FragmentFinancialSystemBinding.inflate(layoutInflater)
 
         setRecyclerView()
 
-        return fragmentSafetyManagementBinding.root
+        return fragmentFinancialSystemBinding.root
     }
     // RecyclerView 설정
     fun setRecyclerView(){
-        fragmentSafetyManagementBinding.recyclerViewSafetyManagement.apply {
+        fragmentFinancialSystemBinding.recyclerViewFinancialSystem.apply {
             // 어댑터 설정
             adapter = OperationSecondRecyclerView(childFragmentManager)
             // 레이아웃
