+++ app/src/main/java/kr/co/lion/application/finalproject_aparttalk/ui/entiremenu/OperationInfo/fragment/RepairPlanFragment.kt
@@ -1,4 +1,4 @@
-package kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptOperationInfo.fragment
+package kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.OperationInfo.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,26 +7,25 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.google.android.material.divider.MaterialDividerItemDecoration
-import kr.co.lion.application.finalproject_aparttalk.R
-import kr.co.lion.application.finalproject_aparttalk.databinding.FragmentBiddingNoticeBinding
-import kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptOperationInfo.adapter.OperationSecondRecyclerView
+import kr.co.lion.application.finalproject_aparttalk.databinding.FragmentRepairPlanBinding
+import kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.OperationInfo.adapter.OperationSecondRecyclerView
 
-class BiddingNoticeFragment : Fragment() {
+class RepairPlanFragment : Fragment() {
 
-    lateinit var fragmentBiddingNoticeBinding: FragmentBiddingNoticeBinding
+    lateinit var fragmentRepairPlanBinding: FragmentRepairPlanBinding
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         // Inflate the layout for this fragment
-        fragmentBiddingNoticeBinding = FragmentBiddingNoticeBinding.inflate(layoutInflater)
+        fragmentRepairPlanBinding = FragmentRepairPlanBinding.inflate(layoutInflater)
 
         setRecyclerView()
 
-        return fragmentBiddingNoticeBinding.root
+        return fragmentRepairPlanBinding.root
     }
 
     // RecyclerView 설정
     fun setRecyclerView(){
-        fragmentBiddingNoticeBinding.recyclerViewBiddingNotice.apply {
+        fragmentRepairPlanBinding.recyclerViewRepairPlan.apply {
             // 어댑터 설정
             adapter = OperationSecondRecyclerView(childFragmentManager)
             // 레이아웃
