+++ app/src/main/java/kr/co/lion/application/finalproject_aparttalk/ui/entiremenu/OperationInfo/fragment/SafetyManagementFragment.kt
@@ -1,4 +1,4 @@
-package kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptOperationInfo.fragment
+package kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.OperationInfo.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,25 +7,24 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.google.android.material.divider.MaterialDividerItemDecoration
-import kr.co.lion.application.finalproject_aparttalk.R
-import kr.co.lion.application.finalproject_aparttalk.databinding.FragmentManagementRegulationBinding
-import kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptOperationInfo.adapter.OperationSecondRecyclerView
+import kr.co.lion.application.finalproject_aparttalk.databinding.FragmentSafetyManagementBinding
+import kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.OperationInfo.adapter.OperationSecondRecyclerView
 
-class ManagementRegulationFragment : Fragment() {
+class SafetyManagementFragment : Fragment() {
 
-    lateinit var fragmentManagementRegulationBinding: FragmentManagementRegulationBinding
+    lateinit var fragmentSafetyManagementBinding: FragmentSafetyManagementBinding
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         // Inflate the layout for this fragment
-        fragmentManagementRegulationBinding = FragmentManagementRegulationBinding.inflate(layoutInflater)
+        fragmentSafetyManagementBinding = FragmentSafetyManagementBinding.inflate(layoutInflater)
 
         setRecyclerView()
 
-        return fragmentManagementRegulationBinding.root
+        return fragmentSafetyManagementBinding.root
     }
     // RecyclerView 설정
     fun setRecyclerView(){
-        fragmentManagementRegulationBinding.recyclerViewManagementRegulation.apply {
+        fragmentSafetyManagementBinding.recyclerViewSafetyManagement.apply {
             // 어댑터 설정
             adapter = OperationSecondRecyclerView(childFragmentManager)
             // 레이아웃
