+++ app/src/main/res/layout/activity_operation_info.xml
@@ -6,12 +6,12 @@
     android:layout_height="match_parent"
     android:background="#FFFFFF"
     android:orientation="vertical"
-    tools:context=".ui.entiremenu.AptOperationInfo.OperationInfoActivity">
+    tools:context=".ui.entiremenu.OperationInfo.OperationInfoActivity">
 
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragmentContainerView"
-        android:name="kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptOperationInfo.fragment.OperationFirstFragment"
+        android:name="kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.OperationInfo.fragment.OperationFirstFragment"
         android:layout_width="match_parent"
         android:layout_height="wrap_content" />
 </LinearLayout>
