+++ app/src/main/res/layout/fragment_bidding_notice.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.entiremenu.AptOperationInfo.fragment.BiddingNoticeFragment">
+    tools:context=".ui.entiremenu.OperationInfo.fragment.BiddingNoticeFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewBiddingNotice"
