+++ app/src/main/res/layout/fragment_billing_statement.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.entiremenu.AptOperationInfo.fragment.BillingStatementFragment" >
+    tools:context=".ui.entiremenu.OperationInfo.fragment.BillingStatementFragment" >
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewBillingStatement"
