+++ app/src/main/res/layout/fragment_contract_info.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.entiremenu.AptOperationInfo.fragment.ContractInfoFragment" >
+    tools:context=".ui.entiremenu.OperationInfo.fragment.ContractInfoFragment" >
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewContractInfo"
