+++ app/src/main/res/layout/fragment_financial_system.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.entiremenu.AptOperationInfo.fragment.FinancialSystemFragment" >
+    tools:context=".ui.entiremenu.OperationInfo.fragment.FinancialSystemFragment" >
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewFinancialSystem"
