+++ app/src/main/res/layout/fragment_management_regulation.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.entiremenu.AptOperationInfo.fragment.ManagementRegulationFragment" >
+    tools:context=".ui.entiremenu.OperationInfo.fragment.ManagementRegulationFragment" >
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewManagementRegulation"
