+++ app/src/main/res/layout/fragment_operation_info_show_bottom_sheet.xml
@@ -6,7 +6,7 @@
     android:background="@drawable/bottom_sheet_rounding"
     android:orientation="vertical"
     android:padding="20dp"
-    tools:context=".ui.entiremenu.AptOperationInfo.fragment.OperationInfoShowBottomSheetFragment">
+    tools:context=".ui.entiremenu.OperationInfo.fragment.OperationInfoShowBottomSheetFragment">
 
     <TextView
         android:id="@+id/textViewOperationInfoShowSubject"
