+++ app/src/main/res/layout/fragment_operation_second.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="#FFFFFF"
-    tools:context=".ui.entiremenu.AptOperationInfo.fragment.OperationSecondFragment">
+    tools:context=".ui.entiremenu.OperationInfo.fragment.OperationSecondFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbarAptOperationInfo"
