+++ app/src/main/res/layout/fragment_repair_plan.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.entiremenu.AptOperationInfo.fragment.RepairPlanFragment" >
+    tools:context=".ui.entiremenu.OperationInfo.fragment.RepairPlanFragment" >
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewRepairPlan"
