+++ app/src/main/res/layout/fragment_safety_management.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.entiremenu.AptOperationInfo.fragment.SafetyManagementFragment" >
+    tools:context=".ui.entiremenu.OperationInfo.fragment.SafetyManagementFragment" >
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewSafetyManagement"
